// src/models/directaRecord.ts
export class DirectaRecord {
    dataOperazione: Date;
    dataValuta: string;
    tipoOperazione: string;
    isin: string;
    ticker: string;
    protocollo: string;
    descrizione: string;
    quantit: number;
    importoEuro: number;
    importoDivisa: number;
    divisa: string;
    riferimentoOrdine: string;
}